Move score handlers to their own module.
